refac rffi into a split module
